Moved the Toaster component to elements directory
