+++ packages/ui-library/src/elements/toaster/style.css
@@ -8,9 +8,5 @@
         .yst-toaster--large {
             @apply yst-w-96;
         }
-
-        .yst-toaster__icon {
-            @apply yst-w-5 yst-h-5;
-        }
     }
 }
